// types/models.ts
export type InvoiceType = "active" | "passive"
// ...
export interface InvoiceKPIs {
  unpaid: number
  paid: number
  overdue: number
  toCollect: number
  collected: number
  overdueAmount: number
}